backend/requirements.txt
alembic==1.5.4
APScheduler==3.7.0
astroid==2.4.2
autopep8==1.5.5
blinker==1.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
Flask==1.1.2
Flask-APScheduler==1.12.2
Flask-Cors==3.0.10
Flask-Migrate==2.6.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
idna==2.10
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonpickle==1.5.1
lazy-object-proxy==1.4.3
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==3.10.0
mccabe==0.6.1
pycodestyle==2.6.0
pylint==2.6.0
PyMySQL==1.0.2
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2021.1
PyYAML==5.4.1
raygun4py==4.3.0
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.23
toml==0.10.2
tzlocal==2.1
urllib3==1.26.3
Werkzeug==1.0.1
wrapt==1.12.1
